Vismach: Merge functions from twp-vismach.py
